refactor(drivers): rename driver modules to driver.py within each package; update dynamic loader and tests
