+++ benchmesh-serial-service/src/benchmesh_service/drivers/tenma_72/__init__.py
@@ -1,2 +1,2 @@
 # Package init for tenma_72 driver
-from .tenma_psu import TenmaPSU as tenma_psu
+from .driver import TenmaPSU as tenma_psu
